Quick changes
I have changed the properties of the project folder so that our main class
that starts with the jar is gui.app.java.
Added function in dom validator, newGUI and Textpanel for letting us
access the active textpanel called from TextPanel.java to NewGui.Java so
that's it's accessible through our FormPanel's actionlistener.

added success confirmation in the gui for when we validate xsd with xml,
in XmlDom.java